refactor(windows): extract PanePreviewRenderer from tab overview
